Progress in open academy and first draft for carpooling
2 != views for the same model ⟶ does not work
