+++ carpooling/security/ir.model.access.csv
@@ -1,2 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_carpooling_carpooling,carpooling.carpooling,model_carpooling_carpooling,base.group_user,1,1,1,1
+access_carpooling_vehicle_trip,carpooling.vehicle_trip,model_carpooling_vehicle_trip,base.group_user,1,1,1,1
