+++ open_academy/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from . import course
-from . import session
+from . import models
+# from . import course
+# from . import session
